ci: Add qodana scan

+++ qodana.ps1
@@ -4,7 +4,9 @@ param (
     [string]
     $Token,
     [string]
-    $Image = "jetbrains/qodana-cdnet:2023.3-eap"
+    $Image = "jetbrains/qodana-cdnet:2023.3-eap",
+    [switch]
+    $SkipScan
 )
 
 $ErrorActionPreference = "Stop"
@@ -18,8 +20,10 @@ if (Test-Path qodana.sln) {
 dotnet new solution -n qodana
 dotnet sln qodana.sln add ./src/OpaDotNet.Wasm/OpaDotNet.Wasm.csproj --in-root
 
-docker run -v ${PWD}:/data/project/ -e QODANA_TOKEN="$token" $Image --solution qodana.sln
+if (-not $SkipScan) {
+    docker run -v ${PWD}:/data/project/ -e QODANA_TOKEN="$token" $Image --solution qodana.sln
 
-if (Test-Path qodana.sln) {
-    rm qodana.sln
+    if (Test-Path qodana.sln) {
+        rm qodana.sln
+    }
 }
